+++ lib/runtime/server/ssr/index.ts
@@ -1,6 +1,6 @@
 export { WatchtowerEvents } from './render-events'
 export { RenderMarkup } from './render-app-to-string'
-export { default as renderPageContents, Assets, WatchtowerOptions } from './full-render'
+export { default as renderPageContents, Assets, ServerSideRenderOptions } from './full-render'
 export {
     ServerRenderResult,
     ServerRenderResultType,